requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
mysqlclient==1.4.6
pyTelegramBotAPI==3.7.2
requests==2.24.0
six==1.15.0
sql==0.4.0
SQLAlchemy==1.3.18
urllib3==1.25.9
